Setup conversion for rework
Moved common models to commons package.

Moved common constants to BaseTable so that one place determines names
and specifiers.

Tables all now have a default start so that if they are used as a link
without a phrase they can have a default starting point.

This is a jump off point, so there are failing tests for the conversions
and other partial work that will get refactored and cleaned up as the
database conversion moves to objects.ˆ