Limit ward count for smaller cities
A Thorp was getting like 20 wards, and a bunch of buildngs so that there
was more buildings than people in the Thorp! It now limits the number of
wards to hopefully keep things a little tighter together. A Thorp had a
few people over a large area of land, so it was generating a lot of
wards for only a few people. Might could have applied population density
to ward size to make the wards consume more space for a sparcer people.

closes #1